tests/Element-getAttribute.html

<!DOCTYPE html>
<title>Element.getAttribute</title>
<script src="testharness.js"></script>
<div id="log"></div>
<script>
var el = document.createElement("div")
document.body.appendChild(el)
test(function() {
  assert_equals(el.getAttribute("bar"), null)
  assert_equals(el.getAttributeNS(null, "bar"), null)
  assert_equals(el.getAttributeNS("", "bar"), null)
}, "Unset attributes return null")
test(function() {
  el.setAttributeNS("ab", "attr", "t1")
  el.setAttributeNS("kl", "attr", "t2")
  assert_equals(el.getAttribute("attr"), "t2")
  assert_equals(el.getAttributeNS("ab", "attr"), "t1")
  assert_equals(el.getAttributeNS("kl", "attr"), "t2")
  assert_equals(el.getAttributeNS(null, "attr"), null)
  assert_equals(el.getAttributeNS("", "attr"), null)
  el.removeAttributeNS("ab", "attr")
  el.removeAttributeNS("kl", "attr")
}, "First set attribute is returned by setAttribute")
test(function() {
  el.setAttribute("style", "color:#fff;")
  assert_equals(el.getAttribute("style"), "color:#fff;")
  assert_equals(el.getAttributeNS(null, "style"), "color:#fff;")
  assert_equals(el.getAttributeNS("", "style"), "color:#fff;")
  el.removeAttribute("style")
}, "Style attributes are not normalized")
test(function() {
  el.setAttributeNS("", "ALIGN", "left")
  assert_equals(el.getAttribute("ALIGN"), null)
  assert_equals(el.getAttribute("align"), null)
  assert_equals(el.getAttributeNS(null, "ALIGN"), "left")
  assert_equals(el.getAttributeNS("", "ALIGN"), "left")
  assert_equals(el.getAttributeNS(null, "align"), null)
  assert_equals(el.getAttributeNS("", "align"), null)
  el.removeAttributeNS("", "ALIGN")
}, "Only lowercase attributes are returned on HTML elements")
test(function() {
  el.setAttributeNS("", "CHEEseCaKe", "tasty")
  assert_equals(el.getAttribute("CHEESECAKE"), null)
  assert_equals(el.getAttribute("CHEEseCaKe"), null)
  assert_equals(el.getAttribute("cheesecake"), null)
  assert_equals(el.getAttributeNS(null, "CHEESECAKE"), null)
  assert_equals(el.getAttributeNS("", "CHEESECAKE"), null)
  assert_equals(el.getAttributeNS(null, "CHEEseCaKe"), "tasty")
  assert_equals(el.getAttributeNS("", "CHEEseCaKe"), "tasty")
  assert_equals(el.getAttributeNS(null, "cheesecake"), null)
  assert_equals(el.getAttributeNS("", "cheesecake"), null)
  el.removeAttributeNS("", "CHEEseCaKe")
}, "Only lowercase attributes are returned on HTML elements")
test(function() {
  el.setAttributeNS("", "align", "left")
  el.setAttributeNS("xx", "align", "right")
  el.setAttributeNS("", "foo", "left")
  el.setAttributeNS("xx", "foo", "right")
  assert_equals(el.getAttribute("align"), "left")
  assert_equals(el.getAttribute("foo"), "left")
  assert_equals(el.getAttributeNS("xx", "align"), "right")
  assert_equals(el.getAttributeNS(null, "foo"), "left")
  assert_equals(el.getAttributeNS("", "foo"), "left")
  el.removeAttributeNS("", "align")
  el.removeAttributeNS("xx", "align")
  el.removeAttributeNS("", "foo")
  el.removeAttributeNS("xx", "foo")
}, "First set attribute is returned with mapped attribute set first")
test(function() {
  el.setAttributeNS("xx", "align", "right")
  el.setAttributeNS("", "align", "left")
  el.setAttributeNS("xx", "foo", "right")
  el.setAttributeNS("", "foo", "left")
  assert_equals(el.getAttribute("align"), "right")
  assert_equals(el.getAttribute("foo"), "right")
  assert_equals(el.getAttributeNS("xx", "align"), "right")
  assert_equals(el.getAttributeNS(null, "foo"), "left")
  assert_equals(el.getAttributeNS("", "foo"), "left")
  el.removeAttributeNS("", "align")
  el.removeAttributeNS("xx", "align")
  el.removeAttributeNS("", "foo")
  el.removeAttributeNS("xx", "foo")
}, "First set attribute is returned with mapped attribute set later")
</script>
